지도 마크업 보이게하고 정보보기 a태그추가

+++ db/migrate/20200501034428_createposts.rb
@@ -11,8 +11,8 @@ class Createposts < ActiveRecord::Migration[5.2]
       t.integer :ImageId
       t.integer :TimeTableId
       t.string :file
-      t.integer :lat
-      t.integer :lng 
+      t.float :lat
+      t.float :lng 
      end 
   end
